requirements.txt
openai
langchain
langchain-openai
pinecone-client
python-dotenv
tiktoken